lang-portal/backend-flask/requirements.txt
pytest>=7.0.0
pytest-flask>=1.2.0flask
flask-cors==5.0.0
invoke
pytest==7.4.3
pytest-flask==1.3.0